feat: now tapo management class is abstact. add impl and mock class to handle development mode without touch camera settings
