+++ docker-compose.yml
@@ -15,6 +15,7 @@ services:
     image: registry.texhome.lan/raspi-home-backend:main
     environment:
       - GPIOZERO_PIN_FACTORY=mock
+      - TAPO_MANAGEMENT=mock
       - TZ=Europe/Rome
     volumes:
       - ./docker/backend/.env:/opt/raspi-home/.env
